// packages/contracts/generated/RunEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Id } from "./Id";
import type { RunEventType } from "./RunEventType";

export type RunEvent = { run_id: Id, seq: number, type: RunEventType, message: string, data: Record<string, unknown> | null, };