// packages/contracts/generated/SuggestionSet.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Change } from "./Change";
import type { Id } from "./Id";
import type { QualityIssue } from "./QualityIssue";
import type { RunContext } from "./RunContext";
import type { SuggestionState } from "./SuggestionState";

export type SuggestionSet = { suggestion_set_id: Id, model_run_id: Id, prediction_id: Id, context: RunContext, changes: Array<Change>, issues: Array<QualityIssue>, score: number | null, state: SuggestionState, };