// packages/contracts/generated/AiPreviewRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AiApprovedGrants } from "./AiApprovedGrants";
import type { StartRunRequest } from "./StartRunRequest";

export type AiPreviewRequest = { request: StartRunRequest, grants: AiApprovedGrants, };